Adding frame_times field to data returned by IPX plugin, and always returning full number of frames in n_frames field.

+++ source/plugins/ipx/ipx.cpp
@@ -1138,7 +1138,7 @@ bool Cipx::loadFrame(int nf)
 #else
     if (fseek(fH, pOffs[nf], SEEK_SET)) {
         fclose(fH);
-        sprintf(errstr, "Can't move file pointer to offset %ld", pOffs[nf]);
+        sprintf(errstr, "Can't move file pointer to offset %ld", (long)pOffs[nf]);
 #endif
         return 0;
     }
